+++ service/web-server/src/auth/routes.ts
@@ -3,22 +3,16 @@
  */
 
 import express from 'express';
-import { AuthController } from './controller';
+import { getAuthController } from './controller';
+import { checkVerification } from './middlewares';
 
 export const getAuthRoutes = () => {
-    const controller = AuthController();
+    const controller = getAuthController();
     const router = express.Router();
-
-    // NOTE: as a temporary solution:
-    //       we fetch and update graknlabsMembers
-    //       in regular intervals in order to
-    //       verify the users against a recent
-    //       list of graknlabs github members.
-    controller.updateGraknlabsMembers();
 
     router.get('/callback', controller.oauthCallback.bind(controller));
 
-    router.get('/verify', controller.checkVerification.bind(controller), controller.verify.bind(controller));
+    router.get('/verify', checkVerification, controller.verify.bind(controller));
 
     return router;
 };
