+++ service/web-server/src/types.ts
@@ -1,0 +1,4 @@
+export interface IGlobal extends NodeJS.Global {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    graknlabsMembers: any[];
+}
